companion-api/requirements.txt
# Core dependencies for Real-Time AI Companion
asyncio
pyaudio
deepgram-sdk>=3.0.0,<4.0.0
openai>=1.0.0
elevenlabs>=1.0.0
chromadb>=0.4.0
numpy>=1.24.0
python-dotenv>=1.0.0
colorama>=0.4.6
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.0.0

